Convert budget section to an over/under grid
