+++ frontend/src/BudgetInput.tsx
@@ -5,15 +5,23 @@ import Stack from "@mui/joy/Stack";
 
 interface Props {
     label: string;
+    isLabelVisible?: boolean;
     value: number;
     onChange: (value: number) => void;
 }
 
-export function BudgetInput({ label, value, onChange }: Props) {
+export function BudgetInput({
+    label,
+    isLabelVisible = true,
+    value,
+    onChange,
+}: Props) {
     return (
-        <Stack direction="row" spacing={1}>
-            <Typography level="h4">{label}</Typography>
+        <Stack spacing={1} maxWidth="100%">
+            {isLabelVisible && <Typography level="body-xs">{label}</Typography>}
             <Input
+                aria-label={label}
+                size="sm"
                 type="number"
                 value={value}
                 onChange={(event) => onChange(Number(event.target.value))}
